Convert runway light vector calculation to SGVec3 and SGQuat

+++ src/Airports/GenAirports850/runway.hxx
@@ -162,7 +162,7 @@ private:
     }
 
     point_list gen_corners( double l_ext, double disp1, double disp2, double w_ext );
-    Point3D gen_runway_light_vector( double angle, bool recip );
+    Point3D gen_runway_light_vector( float angle, bool recip );
     superpoly_list gen_runway_edge_lights( bool recip );
     superpoly_list gen_runway_threshold_lights( const int kind, bool recip );
     superpoly_list gen_runway_center_line_lights( bool recip );
